configure the mac addr via env vars
Up to now, the mac address was being defined on the
network-attachment-definition. This is extremelly inneficient,
since it requires 1 NAD per each pod.

By setting the MAC address via a runtime parameter, indicated in
the pod / vmi network annotation it is possible to have a single
NAD per logical network.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>