Either accept deviceID or master via the stdin args
"DeviceID" is meant to be used when the macvtap device is already
created by the macvtap device plugin, meaning CNI will simply
reconfigure the interface, and move it to the POD's namespace.

"Master" attribute, on the other hand, is meant to be used when
the CNI plugin is responsible for *creating* and configuring the
macvtap interface.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>